Переход на версию Qt6

+++ src/cfield.cpp
@@ -35,7 +35,7 @@ void CField::newGame(int x, int y, int count)
     // Возможно, что типов костяшек не хватит заполнить все поле (есть такие поля)
     // тогда начинаем заполнять оставшиеся поля с начала
     int type{0}, n{0}, max_n(m_count_in_type / 2);
-    int max_type{tiles_manager->tilesNames().size()};
+    int max_type{static_cast<int>(tiles_manager->tilesNames().size())};
     for (int i = 0; i < m_tiles_count; i += 2) {
         m_tiles[v[i]] = type;
         m_tiles[v[i+1]] = type;
